fix: reorganize hexagonal architecture structure
