+++ src/article/application/queries/get-all-articles.handler.ts
@@ -1,5 +1,5 @@
 import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
-import { PaginatedResultType } from '../../../shared/types/paginated-result.type';
+import { PaginatedResultType } from '../../../shared/domain/contracts/paginated-result.type';
 import { GetAllArticleQuery } from './get-all-articles.query';
 import { Article } from '../../domain/entities/article.domain-entity';
 import { ArticleRepository } from '../../domain/repositories/article.repository';
