+++ src/shared/domain/contracts/id-generator.service.ts
@@ -1,0 +1,3 @@
+export interface IdGeneratorService {
+  createId(): string;
+}
