+++ src/shared/domain/contracts/paginated-result.type.ts
