+++ src/shared/infrastructure/services/id-generator.service.impl.ts
@@ -1,0 +1,8 @@
+import { IdGeneratorService } from 'src/shared/domain/contracts/id-generator.service';
+import { v4 as uuidv4 } from 'uuid';
+
+export class IdGeneratorServiceImpl implements IdGeneratorService {
+  createId(): string {
+    return uuidv4();
+  }
+}
