--- src/shared/utils/uuid.ts
@@ -1,3 +1,0 @@
-import { v4 as uuidv4 } from 'uuid';
-
-export const generateUUID = (): string => uuidv4();
